backend/requirements.txt
# Flask and extensions
Flask==3.1.2
flask-cors==6.0.2
flask-restx==1.3.2
flasgger==0.9.7.1
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# MongoDB
pymongo==4.15.5
dnspython==2.8.0

# Data processing
pandas==2.3.3
numpy==2.4.0

# Machine Learning & NLP
scikit-learn==1.8.0
scipy==1.16.3
nltk==3.9.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
vaderSentiment==3.3.2
langdetect==1.0.9

# Visualization
matplotlib==3.10.8
wordcloud==1.9.5

# NLP dependencies
srsly==2.5.2
thinc==8.3.10
cymem==2.0.13
preshed==3.0.12
murmurhash==1.0.15
blis==1.3.3
catalogue==2.0.10
typer-slim==0.20.1
cloudpathlib==0.23.0
confection==0.1.5
weasel==0.4.3
wasabi==1.1.3
smart-open==7.0.0

# Utilities
python-dotenv==1.2.1
python-dateutil==2.9.0.post0
schedule==1.2.2
requests==2.32.5

# Data validation
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# Core dependencies
setuptools==80.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
itsdangerous==2.2.0
click==8.3.1
blinker==1.9.0

# Parsing and formatting
PyYAML==6.0.3
mistune==3.2.0
attrs==25.4.0
annotated-types==0.7.0
pydantic==2.12.5
pydantic-core==2.41.5
typing-extensions==4.15.0
typing-inspection==0.4.2

# Charts and plotting
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pillow==12.0.0
pyparsing==3.3.1

# Utilities
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.2
six==1.17.0
packaging==25.0
joblib==1.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
regex==2025.11.3
wrapt==2.0.1

# Datetime handling
pytz==2025.2
tzdata==2025.3
aniso8601==10.0.1

# Additional utilities
importlib-resources==6.5.2

# Testing
pytest==8.0.0

groq

# RAG (Retrieval-Augmented Generation)
faiss-cpu
sentence-transformers
